# inference: implement an opt-in interface to cache generated sources