@aviatesk 2024-06-27
I've added tests and confirmed that it works well in real-world packages like CassetteOverlay.jl. If there are no objections, I plan to merge it soon.